Refactored FollowerBeaconBlockImporter creation
