+++ app/src/main/kotlin/maru/app/Helpers.kt
@@ -15,52 +15,18 @@
  */
 package maru.app
 
-import java.time.Duration
 import java.util.Optional
 import java.util.UUID
 import kotlin.io.path.Path
 import kotlin.time.Duration.Companion.minutes
 import kotlin.time.toJavaDuration
 import maru.config.ApiEndpointConfig
-import maru.consensus.NewBlockHandler
-import maru.consensus.blockimport.FollowerBeaconBlockImporter
-import maru.consensus.state.FinalizationState
-import maru.executionlayer.client.PragueWeb3JJsonRpcExecutionLayerEngineApiClient
-import maru.executionlayer.manager.ForkChoiceUpdatedResult
-import maru.executionlayer.manager.JsonRpcExecutionLayerManager
 import tech.pegasys.teku.ethereum.executionclient.auth.JwtConfig
 import tech.pegasys.teku.ethereum.executionclient.web3j.Web3JClient
-import tech.pegasys.teku.ethereum.executionclient.web3j.Web3JExecutionEngineClient
 import tech.pegasys.teku.ethereum.executionclient.web3j.Web3jClientBuilder
 import tech.pegasys.teku.infrastructure.time.SystemTimeProvider
 
 object Helpers {
-  fun createFollowerBlockImporter(apiEndpointConfig: ApiEndpointConfig): NewBlockHandler<ForkChoiceUpdatedResult> {
-    val web3JEngineApiClient: Web3JClient =
-      Web3jClientBuilder()
-        .endpoint(apiEndpointConfig.endpoint.toString())
-        .timeout(Duration.ofMinutes(1))
-        .timeProvider(SystemTimeProvider.SYSTEM_TIME_PROVIDER)
-        .executionClientEventsPublisher { }
-        .jwtConfigOpt(wrapJwtPath(apiEndpointConfig.jwtSecretPath))
-        .build()
-    val web3jExecutionLayerClient = Web3JExecutionEngineClient(web3JEngineApiClient)
-    val executionLayerClient = PragueWeb3JJsonRpcExecutionLayerEngineApiClient(web3jExecutionLayerClient)
-    val executionLayerManager =
-      JsonRpcExecutionLayerManager(
-        executionLayerEngineApiClient = executionLayerClient,
-      )
-    return FollowerBeaconBlockImporter(
-      executionLayerManager = executionLayerManager,
-      finalizationStateProvider = {
-        FinalizationState(
-          it.executionPayload.blockHash,
-          it.executionPayload.blockHash,
-        )
-      },
-    )
-  }
-
   private fun wrapJwtPath(jwtPath: String?): Optional<JwtConfig> {
     val jwtConfigPath = Optional.ofNullable(jwtPath)
     return JwtConfig.createIfNeeded(
