+++ consensus/src/main/kotlin/maru/consensus/blockimport/BeaconBlockImporter.kt
@@ -22,11 +22,15 @@ import maru.core.BeaconBlock
 import maru.core.BeaconBlockBody
 import maru.core.BeaconState
 import maru.core.Validator
+import maru.executionlayer.client.PragueWeb3JJsonRpcExecutionLayerEngineApiClient
 import maru.executionlayer.manager.ExecutionLayerManager
 import maru.executionlayer.manager.ForkChoiceUpdatedResult
+import maru.executionlayer.manager.JsonRpcExecutionLayerManager
 import org.apache.logging.log4j.LogManager
 import org.apache.logging.log4j.Logger
 import org.hyperledger.besu.consensus.common.bft.ConsensusRoundIdentifier
+import tech.pegasys.teku.ethereum.executionclient.web3j.Web3JClient
+import tech.pegasys.teku.ethereum.executionclient.web3j.Web3JExecutionEngineClient
 import tech.pegasys.teku.infrastructure.async.SafeFuture
 
 fun interface BeaconBlockImporter {
@@ -40,6 +44,26 @@ class FollowerBeaconBlockImporter(
   private val executionLayerManager: ExecutionLayerManager,
   private val finalizationStateProvider: (BeaconBlockBody) -> FinalizationState,
 ) : NewBlockHandler<ForkChoiceUpdatedResult> {
+  companion object {
+    fun create(web3JEngineApiClient: Web3JClient): NewBlockHandler<ForkChoiceUpdatedResult> {
+      val web3jExecutionLayerClient = Web3JExecutionEngineClient(web3JEngineApiClient)
+      val executionLayerClient = PragueWeb3JJsonRpcExecutionLayerEngineApiClient(web3jExecutionLayerClient)
+      val executionLayerManager =
+        JsonRpcExecutionLayerManager(
+          executionLayerEngineApiClient = executionLayerClient,
+        )
+      return FollowerBeaconBlockImporter(
+        executionLayerManager = executionLayerManager,
+        finalizationStateProvider = {
+          FinalizationState(
+            it.executionPayload.blockHash,
+            it.executionPayload.blockHash,
+          )
+        },
+      )
+    }
+  }
+
   private val log = LogManager.getLogger(this.javaClass)
 
   override fun handleNewBlock(beaconBlock: BeaconBlock): SafeFuture<ForkChoiceUpdatedResult> {
